+++ src/main/java/com/example/cafe/order/domain/Order.java
@@ -58,6 +58,5 @@ public class Order {
 
     public void addOrderItem(OrderItem orderItem) {
         this.orderItems.add(orderItem);
-        orderItem.setOrder(this);
     }
 }