+++ src/test/java/com/example/cafe/member/service/MemberServiceTest.java
@@ -59,7 +59,7 @@ class MemberServiceTest {
 
         assertThatThrownBy(() -> memberService.signup(request))
                 .isInstanceOf(CustomException.class)
-                .hasFieldOrPropertyWithValue("errorCode", ErrorCode.DUPLICATE_USERNAME);
+                .hasFieldOrPropertyWithValue("errorCode", ErrorCode.MEMBER_DUPLICATE_USERNAME);
     }
 
     @Test